fix(lab): int64 coactivation + type hygiene + dead-code cleanup (review pass)
Latent bug fix:
- clustering/sample_membership.py: coactivation matmul cast int32 -> int64
  (X.T @ X silently overflows above ~2.1B samples; same class as the int16
  MergeHistory bug). Downcast to float32 right after is unchanged.

Type hygiene:
- app/backend/state.py: DatasetSearchState now holds typed
  DatasetSearchResult / DatasetSearchMetadata (moved to schemas.py) instead of
  list[dict[str,Any]] / dict[str,Any]; router builds typed objects directly.
- clustering/merge_history.py: meta: dict[str,Any]|None -> typed
  MergeHistoryMeta(origin_path); removes the defensive isinstance(.,Path) scan.

Dead code / honesty:
- app/backend/dependencies.py: drop no-op try/except around StateManager.get().
- harvest/scripts/run_worker_jax.py: one host copy in _to_torch (np.array);
  single-source --batch_size / --activation_threshold defaults from the configs;
  activation_threshold read from method_config (no threaded duplicate).
- harvest/reservoir.py: delete stale TODO on a load-bearing pad mask.
- clustering/merge_history.py: delete dead TODOs + commented-out assert block.
- clustering/merge.py: warnings.warn for expected terminus -> logger.info.

Perf nit:
- app/backend/inference.py: gather next-token probs in JAX before host transfer
  (avoid materializing the full (T, vocab) output_probs to read ~T floats).

+++ param_decomp_lab/clustering/sample_membership.py
@@ -339,7 +339,7 @@ def compute_coactivation_matrix_from_csr(
     component_activity_csr: sparse.csr_matrix,
 ) -> ClusterCoactivationShaped:
     """Compute the full coactivation matrix from a sample-by-component CSR matrix."""
-    activation_matrix = component_activity_csr.astype(np.int32, copy=False)
+    activation_matrix = component_activity_csr.astype(np.int64, copy=False)
     coact = (activation_matrix.T @ activation_matrix).toarray()
     return coact.astype(np.float32, copy=False)
